Refactoring branch

+++ src/main/java/ru/pssbd/fonds/dto/input/ReceiptInput.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 @NoArgsConstructor
@@ -14,6 +15,6 @@ public class ReceiptInput {
     private BigDecimal sum;
 
     @NotNull
-    private Integer fond;
+    private BigInteger fond;
 
 }
